move the header to layout of parkinginfo and booking
