+++ src/app/parking-information/[idSlots]/reserved-parking/[bookingId]/pay-terminate-reserved/page.tsx
@@ -52,7 +52,7 @@ const PayReservedParking: React.FC<{ params: { bookingId: string, idSlots: strin
             try {
                 const response = await terminateBooking(cookieToken); 
                 successAlert("Reserva terminada.");
-                router.push(`/parkings`);
+                router.push(`/payment`);
             } catch (error) {
                 errorAlert((error as Error).message);
             } finally {
